// src/app/interfaces/usuarios.interface.ts
// Generated by https://quicktype.io

//export interface Usuarios {
    //id?:       number;
    //username: string;
    //password: string;
//}

export interface Usuarios {
  id_usuario?: number;  // <-- opcional al registrar
  nombres: string;
  apellidos: string;
  email: string;
  contrasena: string;
  id_rol: number;
  numero_intento?: number;  // <-- opcional
  status?: 'activo' | 'bloqueado'; // <-- opcional
  cedula: string;
  bloqueado_hasta?: string | null; // <-- opcional
}